Move reports under admin
