+++ frontend/constants/routes.ts
@@ -34,10 +34,10 @@ export const FAIR_OFFICER_GUIDE_ROUTE = '/sacfairguide'
 export const GUIDE_ROUTE = (slug?: string): string =>
   slug ? `/guides/${slug}` : '/guides/[page]'
 
-export const REPORT_LIST_ROUTE = '/reports'
-export const REPORT_CREATE_ROUTE = '/reports/create'
+export const REPORT_LIST_ROUTE = '/admin/reports'
+export const REPORT_CREATE_ROUTE = '/admin/reports/create'
 export const REPORT_EDIT_ROUTE = (slug?: number): string =>
-  slug ? `/reports/${slug}` : '/reports/[report]'
+  slug ? `/admin/reports/${slug}` : '/admin/reports/[report]'
 
 export const PROFILE_ROUTE = (slug?: string): string =>
   slug ? `/user/${slug}` : '/user/[user]'
